IHashAlgoInfo 接口增加 DigestLength 属性

+++ HashCalculator/HashAlgos/OfficialImplBlake2bp.cs
@@ -23,13 +23,13 @@ namespace HashCalculator
         [DllImport(Embedded.HashAlgs, CallingConvention = CallingConvention.Cdecl)]
         private static extern int blake2bp_final(IntPtr statePtr, byte[] output, ulong outlen);
 
-        public override ulong MaxOutputSize => 64;
+        public override int MaxOutputSize => 64;
 
         public override string NamePrefix => "BLAKE2bp";
 
         public override AlgoType AlgoGroup => AlgoType.BLAKE2BP;
 
-        public OfficialImplBlake2bp(ulong bitLength) : base(bitLength)
+        public OfficialImplBlake2bp(int bitLength) : base(bitLength)
         {
         }
